Implement Question 3.
To implement Question 3 I considered the isolated tests and the addition
of two integers.

I could consider all kinds of classes that implement the :+ method, but
since I don't have much knowledge of the problem, I would be testing the
implementation of :+ for each of those classes. That's why I chose to do
the isolated version.